Remove Album.get_images, redundant due to refresh.
get_images' sole advantage over refresh is that it returns slightly
less data, ie.e it uses less bandwidth and is easier on imgur.
It's disadvantages is that it put Album into an incosistent state,
because get_images will represent the current state and the other
attributes will represent it's creation time or he last time refresh
was run. For instance, if we get_album and then add images to the
album, and do get_images. Then the number of images returned will be
1 higher than images_count.
Additionally get_images is not needed as it's functionality is
covered by reresh, so it simply adds another method and muddle the
API for no extra functionality