Handle Gallery_images pretending to be Images
For instance http://imgur.com/27dRzvo. Going to the url
gives you what is commonly displayed for a Gallery_image
however, the url structure is identical to non-submitted images.
If the id is used to request information about the image,
the returned data will be in Image format not Gallery_image
despite the id belonging to a submitted image.