Set _has_fetched to False for most albums.
Only for the direct imgur endpoint is information about the
images included. For all methods that return a Album without
using the direct endpoint, _has_fecthed is now set to False
meaning that calls to the images attribute will refresh the
object via the direct endpoint which will set the images
attribute.