Remove anaconda reference
After having this in the code for a bit, I've been feeling a bit
iffy about linking to Binary code I didn't compile or distribute
myself. Doesn't feel great.

A bit sad though since it was a great move to add this to anaconda
and a great initiative to help distribute this package further.
Maybe in the future I could be more involved in that distribution,
which would make me feel more at ease about adding a link to
installation through anaconda here.